Split onboarding into smaller components
